Update README.md
DENIAL OF SERVICE (Using SynFlood) ATTACK 
When two notes are trying to communicate, they may use a THREE WAY HANDSHAKE, the actor will be using a target and attacking machine, the attacking machine will be sending a SYN packet and if the packet goes to the target machine, the target machine will reply with an acknowledgement (SYN/ACK) packet. After that the target machine will be waiting to get the final package from source again which will be ACK packets. The actor will load the target machine with many task and with that it will cause DoS. 
Metaspoilt in kali VM will be used as the attacking maching while centos will be used as the target machine.